refactor(FEAT-DATSET-01): move async connect into AssociatedDatasetService + fix SecurityContext propagation
- Delete redundant AsyncAssociatedDatasetService and its implementation.
  They duplicated the API contract for no abstraction benefit.
- Expose connectDatasetAsync(ConnectDatasetRequest): Mono<...> and
  connectDatasets(List<ConnectDatasetRequest>): Flux<...> directly on
  AssociatedDatasetService, keeping the sync connectDataset method
  intact (matches the precedent where AsyncProjectService owns its
  interface but AssociatedDatasetService just needs one extra
  transport method).
- Fix SecurityContext propagation bug that caused per-record
  AccessDenied on the worker thread. The new methods use
  .contextWrite(ReactiveSecurityContextUtils.reactiveSecurity(securityContext))
  - the exact pattern used by AsyncProjectServiceImpl - so Spring
  Security reactive AOP correctly sets the ThreadLocal before
  @PreAuthorize evaluates on connectDataset.
- Update ConnectDatasetSidebar, AssociatedDatasetsMain, and
  AppConfig to consume AssociatedDatasetService directly.
- Inner records ConnectDatasetRequest and ConnectDatasetResponse
  live on AssociatedDatasetService now.
