+++ datamanager-app/src/main/java/life/qbic/datamanager/AppConfig.java
@@ -26,8 +26,6 @@ import life.qbic.infrastructure.email.EmailServiceProvider;
 import life.qbic.infrastructure.email.identity.IdentityEmailServiceProvider;
 import life.qbic.infrastructure.email.project.ProjectManagementEmailServiceProvider;
 import life.qbic.projectmanagement.application.AppContextProvider;
-import life.qbic.projectmanagement.application.api.AsyncAssociatedDatasetService;
-import life.qbic.projectmanagement.application.api.AsyncAssociatedDatasetServiceImpl;
 import life.qbic.projectmanagement.application.associated_dataset.AssociatedDatasetService;
 import life.qbic.projectmanagement.application.OrganisationRepository;
 import life.qbic.projectmanagement.application.ProjectInformationService;
@@ -324,12 +322,6 @@ public class AppConfig {
     var updateProjectUponQCChange = new UpdateProjectUponQCCreation(
         qualityControlService, projectInformationService, jobScheduler);
     return new QCAddedPolicy(updateProjectUponQCChange);
-  }
-
-  @Bean
-  public AsyncAssociatedDatasetService asyncAssociatedDatasetService(
-      AssociatedDatasetService associatedDatasetService) {
-    return new AsyncAssociatedDatasetServiceImpl(associatedDatasetService);
   }
 
   @Bean
