+++ datamanager-app/src/main/java/life/qbic/datamanager/views/projects/project/datasets/AssociatedDatasetsMain.java
@@ -15,7 +15,6 @@ import life.qbic.datamanager.views.general.Main;
 import life.qbic.datamanager.views.notifications.MessageSourceNotificationFactory;
 import life.qbic.datamanager.views.projects.project.ProjectMainLayout;
 import life.qbic.projectmanagement.application.associated_dataset.AssociatedDatasetService;
-import life.qbic.projectmanagement.application.api.AsyncAssociatedDatasetService;
 import life.qbic.projectmanagement.application.experiment.ExperimentInformationService;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -47,7 +46,6 @@ public class AssociatedDatasetsMain extends Main implements BeforeEnterObserver 
   public static final String PROJECT_ID_ROUTE_PARAMETER = "projectId";
 
   private final transient AssociatedDatasetService associatedDatasetService;
-  private final transient AsyncAssociatedDatasetService asyncAssociatedDatasetService;
   private final transient ExperimentInformationService experimentInformationService;
   private final transient UserPermissions userPermissions;
   private final transient MessageSourceNotificationFactory notificationFactory;
@@ -60,14 +58,11 @@ public class AssociatedDatasetsMain extends Main implements BeforeEnterObserver 
   @Autowired
   public AssociatedDatasetsMain(
       AssociatedDatasetService associatedDatasetService,
-      AsyncAssociatedDatasetService asyncAssociatedDatasetService,
       ExperimentInformationService experimentInformationService,
       UserPermissions userPermissions,
       MessageSourceNotificationFactory notificationFactory) {
     this.associatedDatasetService = requireNonNull(associatedDatasetService,
         "associatedDatasetService must not be null");
-    this.asyncAssociatedDatasetService = requireNonNull(asyncAssociatedDatasetService,
-        "asyncAssociatedDatasetService must not be null");
     this.experimentInformationService = requireNonNull(experimentInformationService,
         "experimentInformationService must not be null");
     this.userPermissions = requireNonNull(userPermissions,
@@ -107,7 +102,6 @@ public class AssociatedDatasetsMain extends Main implements BeforeEnterObserver 
     if (connectDatasetSidebar == null) {
       connectDatasetSidebar = new ConnectDatasetSidebar(
           associatedDatasetService,
-          asyncAssociatedDatasetService,
           experimentInformationService,
           userPermissions,
           notificationFactory);
